pyproject.toml
[project]
name = "minimalist-one-word-subtitle-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20250625",
]
